Inital commit
This comit has four of the clases of my VR shooting game. Theese control some of the main functions regarding the enemyes. Enemy Modifier is an Scriptable Object that alows to easily create and modify new enemyes without adding new prefabs to the project. Attack and Movement controll each one of thease actions. The spawner is in charge of spawning the different enemies in each wave and managing the waves.